Adding restore for print() and moving error handling to browser
Global print() function had previously not been protected and restored like other global tables and functions. Now print will reset like the others do when code is run again.

Error handling is also now moved to browser. This means that pcall and xpcall will no longer be used to catch errors, but the app will be allowed to crash. This will make all types of crashes behave in a similar manner as well as provide more descriptive stack traces for the users.