Minor: bug fix and update knight jump
The _captureListener and removeActiveListeners functions seemed to cause some bugs, but they should be unnecessary anyway, so they are removed to eliminate the bug (and hopefully not cause new ones).

Also adjusting the mouse controls for the knight jump so that the left/right control looks for the player character's position on deciding whether to move left or right instead of the centre of the screen.